pyproject.toml
[project]
name = "sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "langchain>=1.2.12",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.2",
    "openai>=2.28.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.48",
]
